backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.20.0
anthropic==0.28.0
langchain==0.2.5
langchain-anthropic==0.1.15
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
